+++ CHANGELOG.md
@@ -11,6 +11,16 @@ the code was deployed.
 
 ## [unreleased]
 
+### Changed
+
+- Body parsing middleware to match BraveSensor (moved from routes.js to server.js).
+
+### Added
+
+- POST /api/message-clients sends a Twilio text message to all active clients (adapted from BraveSensor) (CU-w9bcb5).
+- Unit tests for the authorize function in api.js (adapted from BraveSensor) (CU-w9bcb5).
+- Integration tests for the /api/message-clients API call (adapted from BraveSensor) (CU-w9bcb5).
+
 ### Security
 
 - Upgrade Chai and brave-alert-lib (CU-8678wgn0p).